backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
rank-bm25
requests
pypdf
beautifulsoup4
python-dotenv
python-multipart